Update the following CLI commands to onlu retrieve fields from the API which are displayed in the CLI:
* st2 execution list
* st2 execution get
* st2 action list
* st2 rule list
* st2 trigger-instance list

If user wants to retrieve and view all the attributes they can use
"--attr all" CLI command argument.
